\name{rsem.ssq}
\alias{rsem.ssq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate the squared sum of a matrix
}
\description{
Calculate the squared sum of a matrix
}
\usage{
rsem.ssq(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A matrix
}
}

\examples{
x<-array(1:6, c(2,3))
rsem.ssq(x)
}

